packages/python-verify/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "botcha-verify"
version = "0.1.0"
description = "Server-side verification middleware for BOTCHA JWT tokens"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Ramin", email = "ramin@dupe.com" }
]
keywords = ["botcha", "jwt", "verification", "middleware", "fastapi", "django", "authentication"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Framework :: Django",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Middleware",
    "Typing :: Typed",
]
dependencies = [
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
fastapi = ["fastapi>=0.100.0"]
django = ["django>=4.2"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "fastapi>=0.100.0",
    "django>=4.2",
]

[project.urls]
Homepage = "https://botcha.ai"
Repository = "https://github.com/dupe-com/botcha"
Documentation = "https://github.com/dupe-com/botcha/tree/main/packages/python-verify"
Issues = "https://github.com/dupe-com/botcha/issues"
